Support favoring specific phrases on Google Text-to-Speech
Set phrases and its boost value (1 to 20) on inspector of GoogleWakewordListener and GoogleVoiceRequestProvidor.

https://cloud.google.com/speech-to-text/docs/reference/rest/v1p1beta1/RecognitionConfig#SpeechContext
