+++ Extension/Google/GoogleSpeechModels.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3380245f54d8f442f89e97bbaea09c0a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
